Make the operation type editable by the adapter
I'm going to start doing mutation queries and need to be able to do
that. Also, could have passed in a boolean for plural or not into the
compiler and let it handle pluralization, but that's less flexible. The
trade off is a slightly more explicit adaper, but that's where I think
any verbosity should exist. The compiler doesn't need to know about
that.