Escape quotes in strings
The GraphQL spec defines a string as being enclosed in double quotes. However, if a string attribute were to have a double quoted character sequence within itself, then we would be faced with GraphQL parsing errors. For example,

```
mutation user_update { user: user_update(id: "1", name: "Jeff "The Jeff" Reine-Adelaide") { id name } }
```

The above results in the server thinking that the string value of the `name` argument is `"Jeff "`, and it promptly complains about the succeeding character `T` which it doesn't expect.